geotiff tests: consolidate VRT missing-sources cluster + _helpers/ foundation (#2393)
PR 1 of the 11-PR epic #2390.

Foundation:
- New _helpers/ with tiff_builders.py (was conftest.py:make_minimal_tiff),
  tiff_surgery.py (relocated from _tiff_surgery.py), markers.py
  (was conftest.py: requires_gpu / requires_loopback / requires_integration
  plus the gpu_available / loopback_available probes).
- conftest.py re-exports the helpers and markers so existing
  `from .conftest import ...` imports keep working. The
  pytest_collection_modifyitems socketserver hack stays in place until
  PR 11.

VRT missing-sources consolidation:
- New vrt/ directory with test_missing_sources.py.
- Folds test_vrt_missing_sources_policy_1799.py (eager-only byte band
  smoke checks) and test_vrt_missing_sources_policy_2367.py
  (eager-plus-chunked float matrix) into one file. Parametrise over the
  eager and dask readers with id="eager" / id="dask"; bad-value matrix
  parametrised over readers and 6 invalid values. No issue numbers in
  test ids.
- test_vrt_missing_sources_default_raise_1843.py stays in place: it
  covers the internal _vrt.read_vrt entry point and the
  XRSPATIAL_GEOTIFF_STRICT env-var override, which is out of the
  consolidated module's surface.

Updated direct importers of _tiff_surgery:
- test_local_tile_byte_cap_1664.py
- test_gpu_tile_byte_cap_2026_05_18.py

Updated release_gate_geotiff.rst to cite the new
vrt/test_missing_sources.py path.

CLUSTER_AUDIT.md maps every old file::test to its new file::test_id;
deleted in a follow-up commit on this branch before merge.

Verification:
- pytest xrspatial/geotiff/tests/vrt/ -v: 22 passed
- pytest xrspatial/geotiff/tests/ -x -q: 5706 passed, 68 skipped,
  6 xfailed, 1 xpassed
