+++ xrspatial/geotiff/tests/_helpers/tiff_surgery.py
