+++ xrspatial/geotiff/tests/vrt/__init__.py
@@ -1,0 +1,1 @@
+"""VRT-focused tests for the GeoTIFF module."""
